+++ rustconn/src/dialogs/settings/mod.rs
@@ -89,6 +89,7 @@ pub struct SettingsDialog {
     copy_on_select_check: adw::SwitchRow,
     show_scrollbar_check: adw::SwitchRow,
     local_shell_command_entry: Entry,
+    close_on_clean_exit_check: adw::SwitchRow,
     // Logging settings
     logging_enabled_row: adw::SwitchRow,
     log_dir_entry: Entry,
@@ -178,6 +179,7 @@ impl SettingsDialog {
             copy_on_select_check,
             show_scrollbar_check,
             local_shell_command_entry,
+            close_on_clean_exit_check,
         ) = create_terminal_page();
 
         let (
@@ -572,6 +574,7 @@ impl SettingsDialog {
             copy_on_select_check,
             show_scrollbar_check,
             local_shell_command_entry,
+            close_on_clean_exit_check,
             logging_enabled_row,
             log_dir_entry,
             retention_spin,
@@ -870,6 +873,7 @@ impl SettingsDialog {
             &self.copy_on_select_check,
             &self.show_scrollbar_check,
             &self.local_shell_command_entry,
+            &self.close_on_clean_exit_check,
             &settings.terminal,
         );
 
@@ -987,6 +991,7 @@ impl SettingsDialog {
         let copy_on_select_check_clone = self.copy_on_select_check.clone();
         let show_scrollbar_check_clone = self.show_scrollbar_check.clone();
         let local_shell_command_entry_clone = self.local_shell_command_entry.clone();
+        let close_on_clean_exit_check_clone = self.close_on_clean_exit_check.clone();
 
         // Logging controls
         let logging_enabled_row_clone = self.logging_enabled_row.clone();
@@ -1091,6 +1096,7 @@ impl SettingsDialog {
                 &copy_on_select_check_clone,
                 &show_scrollbar_check_clone,
                 &local_shell_command_entry_clone,
+                &close_on_clean_exit_check_clone,
                 log_timestamps_check_clone.is_active(),
             );